SAK-16034 all unit test now working again
git-svn-id: https://source.sakaiproject.org/svn/search/trunk@67756 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ search/search-util/src/java/org/sakaiproject/search/util/FileUtils.java
@@ -81,10 +81,12 @@ public class FileUtils
 		{
 			public void doFile(File file) throws IOException
 			{
-				file.delete();
-				if (file.exists())
-				{
-					throw new IOException("Failed to delete  " + file.getPath());
+				if (file != null) {
+					file.delete();
+					if (file.exists())
+					{
+						throw new IOException("Failed to delete  " + file.getPath());
+					}
 				}
 			}
 
@@ -101,7 +103,7 @@ public class FileUtils
 	public static void recurse(File f, RecurseAction action) throws IOException
 	{
 		action.doBeforeFile(f);
-		if (f.isDirectory())
+		if (f !=null && f.isDirectory())
 		{
 			File[] files = f.listFiles();
 			if (files != null)
